Fix multiple operator issue and slice = sign
Added a condition to remove the = sign in the
running calculation if further calculations are
done after equals but before clear().

Also noticed that repeated clicks on the operators
continued to operate and update sum even if no new
numbers are entered by user. I prevented this
by adding a return if the display is empty.