Add if statement to number listener
Known issue from previous commit. If user equated
a calculation with '= 'calculation could be extended
if new operator was then chosen. However, if new
operator was not chosen and instead number was selected
old equation would remain in the upper display.

This commit clears the upper display and stored variables
from the users previous calculation so that if numbers
are selected after '=' a new calculation can begin.